printing partitioning summary

+++ niftynet/io/image_sets_partitioner.py
@@ -24,36 +24,55 @@ from niftynet.utilities.util_csv import match_and_write_filenames_to_csv
 from niftynet.utilities.util_csv import write_csv
 
 COLUMN_UNIQ_ID = 'subject_id'
+COLUMN_PHASE = 'phase'
 TRAIN_ID = 'train'
 VALID_ID = 'validation'
 INFER_ID = 'inference'
-SUPPORTED_PHASES = {TRAIN_ID, VALID_ID, INFER_ID}
+ALL_ID = 'all'
+SUPPORTED_PHASES = {TRAIN_ID, VALID_ID, INFER_ID, ALL_ID}
 
 
 @singleton
 class ImageSetsPartitioner(object):
-    def __init__(self,
-                 data_param,
-                 dataset_split_file="",
-                 ratios=(0.1, 0.1)):
+    _file_list = None
+    _partition_ids = None
+    data_param = None
+    ratios = None
+
+    # for saving the splitting index
+    dataset_split_file = ""
+    # default location for searching the image files
+    default_image_file_location = \
+        NiftyNetGlobalConfig().get_niftynet_home_folder()
+
+    def initialise(self,
+                   data_param,
+                   new_partition=True,
+                   data_split_file="./test.csv",
+                   ratios=(0.1, 0.1)):
         self.data_param = data_param
+        self.dataset_split_file = data_split_file
         self.ratios = ratios
-        self.dataset_split_file = dataset_split_file
-        self.default_image_file_location = \
-            NiftyNetGlobalConfig().get_niftynet_home_folder()
-        self._file_list = None
-
-    def initialise(self):
         self.load_data_sections_by_subject()
-        self.split_dataset()
+        self.randomly_split_dataset(overwrite=new_partition)
 
-    @property
-    def number_of_subjects(self):
-        return self._file_list.shape[0].count()
+    def number_of_subjects(self, phase='all'):
+        phase = look_up_operations(phase.lower(), SUPPORTED_PHASES)
+        if phase == ALL_ID:
+            return self._file_list[COLUMN_UNIQ_ID].count()
+        selector = self._partition_ids[COLUMN_PHASE] == phase
+        return self._partition_ids[selector].count()[COLUMN_UNIQ_ID]
 
     def get_file_list(self, phase='train'):
         phase = look_up_operations(phase.lower(), SUPPORTED_PHASES)
-        pass
+        if phase == ALL_ID:
+            self._file_list = self._file_list.sort_index()
+            return self._file_list
+        selector = self._partition_ids[COLUMN_PHASE] == phase
+        selected = self._partition_ids[selector][[COLUMN_UNIQ_ID]]
+        if selected.empty:
+            return []
+        return pandas.merge(self._file_list, selected, on=COLUMN_UNIQ_ID)
 
     def load_data_sections_by_subject(self):
         if not self.data_param:
@@ -67,11 +86,11 @@ class ImageSetsPartitioner(object):
                 # adding all rows of the first modality
                 self._file_list = modality_file_list
                 continue
-            n_rows = self._file_list.shape[0].count()
+            n_rows = self._file_list[COLUMN_UNIQ_ID].count()
             self._file_list = pandas.merge(self._file_list,
                                            modality_file_list,
                                            on=COLUMN_UNIQ_ID)
-            if self._file_list.shape[0].count() < n_rows:
+            if self._file_list[COLUMN_UNIQ_ID].count() < n_rows:
                 tf.logging.warning('rows not matched in section [%s]',
                                    section_name)
 
@@ -141,14 +160,45 @@ class ImageSetsPartitioner(object):
             raise
         return csv_list
 
-    def split_dataset(self):
-        validation_fraction, inference_fraction = self.ratios
-        count_valid = math.ceil(self.number_of_subjects * validation_fraction)
-        count_infer = math.ceil(self.number_of_subjects * inference_fraction)
-        count_train = self.number_of_subjects - count_infer - count_valid
-        phases = [TRAIN_ID] * count_train + \
-                 [VALID_ID] * count_valid + \
-                 [INFER_ID] * count_infer
-        random.shuffle(phases)
-        write_csv(self.dataset_split_file,
-                  zip(self._file_list[COLUMN_UNIQ_ID], phases))
+    def randomly_split_dataset(self, overwrite=False):
+        if overwrite:
+            validation_fraction, inference_fraction = self.ratios
+            n_total = self.number_of_subjects()
+            n_valid = int(math.ceil(n_total * validation_fraction))
+            n_infer = int(math.ceil(n_total * inference_fraction))
+            n_train = int(n_total - n_infer - n_valid)
+            phases = [TRAIN_ID] * n_train + \
+                     [VALID_ID] * n_valid + \
+                     [INFER_ID] * n_infer
+            random.shuffle(phases)
+            write_csv(self.dataset_split_file,
+                      zip(self._file_list[COLUMN_UNIQ_ID], phases))
+
+        if os.path.isfile(self.dataset_split_file):
+            self._partition_ids = pandas.read_csv(
+                self.dataset_split_file,
+                header=None,
+                names=[COLUMN_UNIQ_ID, COLUMN_PHASE])
+            return
+        tf.logging.fatal('Unable to load partition file from %s',
+                         self.dataset_split_file)
+        raise ValueError
+
+    def __str__(self):
+        return self.to_string()
+
+    def to_string(self):
+        summary_str = '\nNumber of subjects {}, '.format(
+            self.number_of_subjects())
+        if self.ratios:
+            summary_str += 'data partitioning (percentage): \n' \
+                           '-- {} {} ({}),\n' \
+                           '-- {} {} ({}), \n' \
+                           '-- {} {}.\n'.format(
+                VALID_ID, self.number_of_subjects(VALID_ID), self.ratios[0],
+                INFER_ID, self.number_of_subjects(INFER_ID), self.ratios[1],
+                TRAIN_ID, self.number_of_subjects(TRAIN_ID))
+        else:
+            summary_str += '-- using all subjects ' \
+                           '(without data partitioning).\n'
+        return summary_str
